Extracted the default sounds
