+++ planet/Objects.ocd/Clonk.ocd/Sounds.ocd/DefCore.txt
@@ -1,0 +1,5 @@
+[DefCore]
+id=Clonk_Sounds
+Version=8,0
+Category=C4D_StaticBack
+HideInCreator=true
